Finalizando ultimo desafio do curso: Usando bigDecimal e o metodo round para arredondar pontos flutuantes.

+++ src/main/kotlin/br/com/alura/alugames/modelo/IRecomenda.kt
@@ -1,6 +1,8 @@
 package br.com.alura.alugames.modelo
 
+import java.math.BigDecimal
+
 interface IRecomenda {
-    val media: Double
+    val media: BigDecimal
 
     fun recomendar(nota: Int)}
